pyproject.toml
[project]
name = "poc-lineage-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-sdk>=0.88.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.1",
]
